scripts/datenschleuder_agent/requirements.txt
redis==5.0.1
python-dotenv==1.0.0
pyyaml>=6.0
